refactor: 각 도메인별 디렉토리 변경
