+++ src/main/java/christmas/domain/date/service/DateService.java
@@ -1,6 +1,6 @@
-package christmas.service;
+package christmas.domain.date.service;
 
-import christmas.domain.date.VisitDate;
+import christmas.domain.date.domain.VisitDate;
 
 public class DateService {
