+++ src/main/java/christmas/domain/discount/policy/Discount.java
@@ -1,7 +1,7 @@
 package christmas.domain.discount.policy;
 
 
-import christmas.domain.order.Order;
+import christmas.domain.order.domain.Order;
 
 public interface Discount {
     boolean isSupport(Order order);
