+++ src/main/java/christmas/service/RestaurantService.java
@@ -1,12 +1,15 @@
 package christmas.service;
 
-import christmas.domain.date.VisitDate;
-import christmas.domain.discount.DiscountDetails;
-import christmas.domain.order.Order;
-import christmas.domain.order.OrderMenu;
-import christmas.dto.DiscountDetailsDto;
-import christmas.dto.OrderMenuDto;
-import christmas.dto.VisitDateDto;
+import christmas.domain.date.service.DateService;
+import christmas.domain.date.domain.VisitDate;
+import christmas.domain.discount.domain.DiscountDetails;
+import christmas.domain.discount.service.DiscountService;
+import christmas.domain.order.domain.Order;
+import christmas.domain.order.domain.OrderMenu;
+import christmas.domain.order.service.OrderMenuService;
+import christmas.domain.discount.dto.DiscountDetailsDto;
+import christmas.domain.order.dto.OrderMenuDto;
+import christmas.domain.date.dto.VisitDateDto;
 
 public class RestaurantService {
