+++ src/test/java/christmas/domain/discount/ChristmasDiscountTest.java
@@ -3,10 +3,10 @@ package christmas.domain.discount;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
-import christmas.domain.date.VisitDate;
+import christmas.domain.date.domain.VisitDate;
 import christmas.domain.discount.policy.ChristmasDiscount;
-import christmas.domain.order.Order;
-import christmas.domain.order.OrderMenu;
+import christmas.domain.order.domain.Order;
+import christmas.domain.order.domain.OrderMenu;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
