+++ src/test/java/christmas/domain/discount/EventBadgeTest.java
@@ -3,7 +3,7 @@ package christmas.domain.discount;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
-import christmas.domain.discount.event.EventBadge;
+import christmas.domain.discount.enums.EventBadge;
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
