+++ src/test/java/christmas/domain/menu/MenuTest.java
@@ -2,7 +2,7 @@ package christmas.domain.menu;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import christmas.domain.order.menu.Menu;
+import christmas.domain.order.enums.Menu;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
