+++ src/test/java/christmas/domain/order/OrderMenuTest.java
@@ -3,6 +3,7 @@ package christmas.domain.order;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
+import christmas.domain.order.domain.OrderMenu;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
