+++ src/test/java/christmas/domain/order/OrderTest.java
@@ -4,8 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
-import christmas.domain.date.VisitDate;
-import christmas.domain.discount.event.EventDate;
+import christmas.domain.date.domain.VisitDate;
+import christmas.domain.discount.enums.EventDate;
+import christmas.domain.order.domain.Order;
+import christmas.domain.order.domain.OrderMenu;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
