+++ src/test/java/christmas/domain/service/DiscountServiceTest.java
@@ -3,16 +3,16 @@ package christmas.domain.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
-import christmas.domain.date.VisitDate;
-import christmas.domain.discount.DiscountDetails;
+import christmas.domain.date.domain.VisitDate;
+import christmas.domain.discount.domain.DiscountDetails;
 import christmas.domain.discount.policy.ChristmasDiscount;
 import christmas.domain.discount.policy.GiftDiscount;
 import christmas.domain.discount.policy.StarDayDiscount;
 import christmas.domain.discount.policy.WeekDayDiscount;
 import christmas.domain.discount.policy.WeekendDiscount;
-import christmas.domain.order.Order;
-import christmas.domain.order.OrderMenu;
-import christmas.service.DiscountService;
+import christmas.domain.order.domain.Order;
+import christmas.domain.order.domain.OrderMenu;
+import christmas.domain.discount.service.DiscountService;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
