+++ src/test/java/christmas/domain/service/RestaurantServiceTest.java
@@ -7,9 +7,9 @@ import christmas.domain.discount.policy.GiftDiscount;
 import christmas.domain.discount.policy.StarDayDiscount;
 import christmas.domain.discount.policy.WeekDayDiscount;
 import christmas.domain.discount.policy.WeekendDiscount;
-import christmas.service.DateService;
-import christmas.service.DiscountService;
-import christmas.service.OrderMenuService;
+import christmas.domain.date.service.DateService;
+import christmas.domain.discount.service.DiscountService;
+import christmas.domain.order.service.OrderMenuService;
 import christmas.service.RestaurantService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
